improved user experience and added animations

+++ To-Do List/To-Do List/View/ListView.swift
@@ -11,19 +11,26 @@ struct ListView: View {
 
     @EnvironmentObject var listViewModel: ListViewModel
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("To-Do List 📝")
         .navigationBarItems(leading: EditButton(), trailing:
             NavigationLink("Add", destination: AddView())
@@ -32,7 +39,6 @@ struct ListView: View {
 }
 
 struct ListView_Previews: PreviewProvider {
-    //1.
     static var previews: some View {
         NavigationView {
             ListView()
